Remove passive replication scenario tests
