+++ tests/passivepath/control_test.go
@@ -8,12 +8,24 @@ import (
 	"github.com/stretchr/testify/require"
 	sdkclient "go.temporal.io/sdk/client"
 	sdkworker "go.temporal.io/sdk/worker"
+	"go.temporal.io/sdk/workflow"
 	"go.temporal.io/server/common/log"
 )
 
-// TestControl_NoHook is the baseline: the same cluster config and the same workflows,
-// with the passive-path hook NOT installed. Any log noise that appears here is
-// pre-existing and not attributable to the diversion.
+func controlActivity(_ context.Context, value string) (string, error) {
+	return value, nil
+}
+
+func controlWorkflow(ctx workflow.Context) (string, error) {
+	ctx = workflow.WithActivityOptions(ctx, workflow.ActivityOptions{
+		ScheduleToCloseTimeout: 20 * time.Second,
+	})
+	var result string
+	err := workflow.ExecuteActivity(ctx, controlActivity, "ok").Get(ctx, &result)
+	return result, err
+}
+
+// TestControl_NoHook verifies that the cluster setup works without the hook installed.
 func TestControl_NoHook(t *testing.T) {
 	sdkworker.SetStickyWorkflowCacheSize(0)
 	logger := log.NewTestLogger()
@@ -31,9 +43,8 @@ func TestControl_NoHook(t *testing.T) {
 
 	taskQueue := "control-tq"
 	w := sdkworker.New(sdkClient, taskQueue, sdkworker.Options{})
-	w.RegisterWorkflow(passivePathWorkflow)
-	w.RegisterWorkflow(parallelActivityWorkflow)
-	w.RegisterActivity(echoActivity)
+	w.RegisterWorkflow(controlWorkflow)
+	w.RegisterActivity(controlActivity)
 	require.NoError(t, w.Start())
 	t.Cleanup(w.Stop)
 
@@ -42,16 +53,9 @@ func TestControl_NoHook(t *testing.T) {
 
 	run, err := sdkClient.ExecuteWorkflow(ctx, sdkclient.StartWorkflowOptions{
 		ID: "control-basic", TaskQueue: taskQueue, WorkflowRunTimeout: 45 * time.Second,
-	}, passivePathWorkflow, false)
+	}, controlWorkflow)
 	require.NoError(t, err)
 	var s string
 	require.NoError(t, run.Get(ctx, &s))
-
-	run2, err := sdkClient.ExecuteWorkflow(ctx, sdkclient.StartWorkflowOptions{
-		ID: "control-parallel", TaskQueue: taskQueue, WorkflowRunTimeout: 45 * time.Second,
-	}, parallelActivityWorkflow)
-	require.NoError(t, err)
-	var n int
-	require.NoError(t, run2.Get(ctx, &n))
-	require.Equal(t, parallelActivityCount, n)
+	require.Equal(t, "ok", s)
 }
